+++ notebook/original/BagofVisualWords_python3x.py
@@ -7,6 +7,9 @@ from sklearn.neighbors import KNeighborsClassifier
 from sklearn.metrics import confusion_matrix
 from matplotlib import pyplot as plt
 import itertools
+from sklearn import svm
+
+from visualization import plot_accuracy_vs_time, plot_confusion_matrix
 
 def open_pkl(pkl_file):
     """
@@ -16,70 +19,25 @@ def open_pkl(pkl_file):
         data = pickle.load(f)    
     return data
 
-def plot_accuracy_vs_time(x, y1, y2, feature_name, title):
-    """
-    This function plots a doble axis figure.
-    Feature name and title can be modified to be plot
-    """    
-    fig, ax1 =plt.subplots()
-    color = 'tab:red'
-    ax1.set_xlabel(feature_name)
-    ax1.set_ylabel('accuracy', color=color)
-    ax1.plot(x, y1, color=color)
-    ax1.tick_params(axis='y', labelcolor=color)
-    
-    ax2 = ax1.twinx()  # instantiate a second axes that shares the same x-axis
-    color = 'tab:blue'
-    ax2.set_ylabel('time (s)', color=color)  
-    ax2.plot(x,y2, color=color)
-    ax2.tick_params(axis='y', labelcolor=color)
-    
-    fig.tight_layout()  # otherwise the right y-label is slightly clipped
-    plt.title(title)
-    plt.show()     
-
-def plot_confusion_matrix(cm, classes, normalize=False, 
-                          title='Confusion matrix', cmap=plt.cm.Blues):
-    """
-    This function prints and plots the confusion matrix.
-    Normalization can be applied by setting `normalize=True`.
-    """
-    if normalize:
-        cm = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
-
-    plt.figure()
-    plt.imshow(cm, interpolation='nearest', cmap=cmap)
-    plt.title(title)
-    plt.colorbar()
-    tick_marks = np.arange(len(classes))
-    plt.xticks(tick_marks, classes, rotation=45)
-    plt.yticks(tick_marks, classes)
-
-    fmt = '.2f' if normalize else 'd'
-    thresh = cm.max() / 2.
-    for i, j in itertools.product(range(cm.shape[0]), range(cm.shape[1])):
-        plt.text(j, i, format(cm[i, j], fmt),
-                 horizontalalignment="center",
-                 color="white" if cm[i, j] > thresh else "black")
-
-    plt.ylabel('True label')
-    plt.xlabel('Predicted label')
-    plt.tight_layout()
-    plt.show()
 
 
 def compute_detector(sift_step_size, sift_scale, n_features = 300):
     """
-    Computes Sift detector objest.
+    Computes Sift detector object.
     Computes mesh of KPs using a custom step size and scale value(s).
     """    
     SIFTdetector = cv2.xfeatures2d.SIFT_create(nfeatures=n_features)
+    
     if(isinstance(sift_scale, list) == False):
         sift_scale = [sift_scale]
+    
     kpt = [cv2.KeyPoint(x, y, scale) for y in 
            range(0, 256, sift_step_size) for x in 
            range(0, 256, sift_step_size) for scale in sift_scale ]
+    
+    
     return (SIFTdetector, kpt)    
+
 
 def create_BOW(dense, SIFTdetector, kpt, k_codebook):
     
@@ -116,17 +74,30 @@ def create_BOW(dense, SIFTdetector, kpt, k_codebook):
         visual_words[i,:] = np.bincount(words, minlength = k_codebook)
         #norm here
     return codebook, visual_words, train_labels 
+
+
+def init_classifier(type_classifier, knn_param, svm_param):
+    
+    if type_classifier == "KNN":
+        
+        return KNeighborsClassifier(n_neighbors=knn_param["k_classifier"], n_jobs=-1, 
+                               metric=knn_param["distance_method"])
+        
+    elif type_classifier == "SVM":
+        return svm.SVC(kernel = svm_param["kernel"], C = svm_param["C"])
+    
+    else:
+        print ("Invalid Classifier")
+        return 0
     
 def classify_BOW(dense, k_codebook, visual_words, codebook, train_labels, 
-                 k_classifier, distance_method):
+                 clf):
     
     test_images_filenames = open_pkl('test_images_filenames.dat')
     test_labels = open_pkl('test_labels.dat')
-          
-    knn = KNeighborsClassifier(n_neighbors=k_classifier, n_jobs=-1, 
-                               metric=distance_method)
-
-    knn.fit(visual_words, train_labels) 
+    
+      
+    clf.fit(visual_words, train_labels) 
     
     visual_words_test=np.zeros((len(test_images_filenames),k_codebook),
                                dtype=np.float32)
@@ -142,13 +113,15 @@ def classify_BOW(dense, k_codebook, visual_words, codebook, train_labels,
         else:
             (kp, des) = SIFTdetector.detectAndCompute(gray,None)  
             
-        words=codebook.predict(des)
-        visual_words_test[i,:]=np.bincount(words,minlength=k_codebook)
+        words = codebook.predict(des)
+        visual_words_test[i,:] = np.bincount(words,minlength=k_codebook)
         
 
 
-    accuracy = 100*knn.score(visual_words_test, test_labels)
-    predicted_labels = knn.predict(visual_words_test)
+    accuracy = 100*clf.score(visual_words_test, test_labels)
+    
+    predicted_labels = clf.predict(visual_words_test)
+    
     unique_labels = list(set(train_labels))
 
     # Compute confusion matrix
@@ -166,12 +139,21 @@ if __name__ == "__main__":
     sift_scale = [8,16,32]
     # Dense/Normal Sift 
     dense = True
+    
     # Number of clusters in KMeans, size of codebook (words)
     k_codebook = 128
-    # number of neightbours taken into account for the classifier
-    k_classifier =  5        
-    # Distance metric use to match 
-    distance_method = "euclidean"
+    
+    type_classifier = "KNN" # SVM
+    
+    ## Values for the classifiers
+    knn_dict =	{
+      "k_classifier": 5,
+      "distance_method": "euclidean",
+    }
+    svm_dict =	{
+      "kernel": "linear",
+      "C": 1,
+    }
 
     accuracy_list = []    
     time_list = []
@@ -179,26 +161,37 @@ if __name__ == "__main__":
     
     #range_value = [int(2**(e)) for e in range(3,6)]
     range_value = [[16],[16,32],[8,16,32],[8,16,32,64]]
+    range_value = sift_scale
     
     for sift_scale in range_value:
         start = time.time()   
+        
         (SIFTdetector, kpt) = compute_detector(sift_step_size, sift_scale)
         print(len(kpt))
+        
         codebook, visual_words, labels = create_BOW(dense, SIFTdetector, 
                                                     kpt, k_codebook)   
         bow_time = time.time()
+        
+        
+        classifier = init_classifier(type_classifier, knn_dict, svm_dict )
+        
+        
         accuracy, cnf_matrix, unique_labels = classify_BOW(dense, k_codebook, 
                                                            visual_words, codebook, 
-                                                           labels, k_classifier, 
-                                                           distance_method)
+                                                           labels, classifier)
         accuracy_list.append(accuracy)
         class_time = time.time()
         ttime = class_time-start
         time_list.append(ttime)
+        
         print ("Accuracy:",accuracy,"Total Time: ", class_time-start,
                ". BOW Time: ", bow_time-start,
                ". Classification Time: ", class_time-bow_time) 
  
+    
+    
+    # Plots
     range_value = [1,2,3,4]
     plot_accuracy_vs_time(range_value, accuracy_list, time_list, 
                        feature_name = 'Number of SIFT scales', title = "DSIFT")
